[Mod] move all bybit codes into bybit_gateway.py
